bgpd: Change reference bandwidth for UCMP
Modify the reference bandwidth used to compute the BGP link-bandwidth
extended community value from 1 Gbps to 1 Mbps. This is to ensure that
if the destination (anycast IP) is present in many places and there is
a large amount of parallel paths, the computed/cumulative bandwidths
still stay within range.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-28814