[CUMULUS] zebra: remove proto only NHG in show run
Remove `zebra nexthops proto only` from show run since it's
enabled by default in cumulus FRR. Having it in show run makes
it exist in customer configs which will become an issue when
they upgrade to a version that does not have it enabled by default
and they retian that config. The idea here is to prevent a future bug
by masking this in the short term.

Signed-off-by: Stephen Worley <sworley@cumulusnetworks.com>