zebra: display AdvRetransTimer in mismatch log
Currently we display a log that indicates we've received an RA whose
AdvRetransTimer does not match our own, however we don't actually log
what our configured value is or what value we got from the packet.
Let's include this in the log so an operator will actually know what's
going on.

before:
```

```

after:
```

```

Signed-off-by: Trey Aspelund <taspelund@nvidia.com>